+++ app/tailwind.css
@@ -74,3 +74,7 @@
     @apply bg-background text-foreground;
   }
 }
+
+.container {
+    @apply flex justify-between p-10 w-full max-w-7xl
+}
