+++ package.json
@@ -21,6 +21,7 @@
     "lucide-react": "^0.403.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "simple-icons": "^13.1.0",
     "tailwind-merge": "^2.4.0",
     "tailwindcss-animate": "^1.0.7"
   },
